feature: SubMenu, Comment Form and New Post Form
Upped the state from HomePage to App Component, created the SubItem component to deal with submenus. Created the Comment, CommentList, CommentForm and CommentSection components and finished the fields on the AddPostModal Component. Changed the way the vote button works from going from thumbs up to thumbs down, to just changing colors and the text besides the icon. Changed the "comments" button to a "Join the Discussion" button that will open the Post Details Page, probably in the next commit.